+++ spec/social_buttons/config_spec.rb
@@ -1,0 +1,13 @@
+require 'spec_helper'
+
+describe SocialButtons do
+  describe '#config' do
+    pending
+  end
+
+  SocialButtons.names.each do |name|
+    describe ".#{name}" do
+      pending
+    end
+  end
+end
